[FIX] Rewrite delayed queue
Delayed queue is now shorter and more efficient, only doing process_rebuild of validation steps.
Features are now able to rebuild arch and arch_eval step on the fly before answering the request.
This way index is always matching the ropes, and arch is valid too.
Validation is delayed to inactivity period, but it is not blocking requests.
Validation (arch and arch_eval) of functions can be done on the fly too