[FIX] follow_ref dropping evaluations incorrectly
The issue:
In follow_ref's main loop, popping from the front of the results Deque
when meaning to replace an evaluation would not necessarily remove the
correct one.

In order to fix this and reduce complexity of the main loop with a
moving index over a Deque, we switch to a work queue approach:
- evaluations that need to be processed are pushed to and popped from
the work queue,
- final evaluations are pushed to the results vector.

Other changes:
- `acc` was renamed to `visited` for clarity.
- some indentation could be reduced with `let else` constructs, improving
readability.