[FIX] correctly identify non-registry model classes
We are trying to do that by checking _register = False.
The check was failing because we handled only one evaluation.
However, in latest versions we also have a type hint, which is its own evaluation.
For that, we implement here a smarter search that only looks are evaluations with values.
In that case, we should be more generic in catching cases.
Consider using this as the check, instead of checking symbol trees